Solved circular dependency problems.
